requirements.txt
async-exit-stack==1.0.1
async-generator==1.10
cffi==1.14.3
click==7.1.2
cryptography==2.9.2
dataclasses==0.7
fastapi==0.61.1
h11==0.10.0
pbkdf2==1.3
pycparser==2.20
pydantic==1.6.1
PyJWT==1.7.1
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.12.1
